'add aapb record block handling, remove aapb record group field handling'
